requirements.txt
# cdk stuff
aws-cdk-lib
constructs
python-dotenv

# lambda stuff
boto3
awscli
requests
pytest

# analysis stuff
numpy
pandas
matplotlib
seaborn

# dev stuff
jupyter
pylint
rope
black
isort
